extract usefullscreen
